requirements.txt
torch~=2.7.0+cu128
torchvision==0.17.2
flash_attn-2.7.3+cu11torch2.2cxx11abiFALSE-cp311-cp311-linux_x86_64.whl
timm==1.0.14
albumentations==2.0.4
onnx==1.14.0
onnxruntime==1.15.1
pycocotools==2.0.7
PyYAML~=6.0.2
scipy~=1.15.3
onnxslim==0.1.31
onnxruntime-gpu==1.18.0
gradio==4.44.1
opencv-python~=4.11.0.86
psutil~=7.0.0
py-cpuinfo==9.0.0
huggingface-hub~=0.32.4
safetensors==0.4.3
numpy~=2.2.6
supervision==0.22.0
ultralytics~=8.3.63
pillow~=10.4.0
matplotlib~=3.10.3
ultralytics-thop~=2.0.14
requests~=2.32.3
tqdm~=4.67.1
ttkbootstrap~=1.13.10